Accreditation implementation

+++ java-challenge/src/main/java/com/project/java_challenge/repositories/AccreditationRepository.java
@@ -1,9 +1,8 @@
 package com.project.java_challenge.repositories;
-/**
+
 import com.project.java_challenge.entities.Accreditation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AccreditationRepository extends JpaRepository<Accreditation, Long> {
+public interface AccreditationRepository extends JpaRepository<Accreditation, Integer> {
 
 }
- */
